+++ extensions/bitwarden/src/utils/cards.ts
@@ -27,7 +27,7 @@ const getCardValue = (key: string, value: string) => {
 
 export function getCardDetailsMarkdown(itemName: string, card: Card) {
   return `# 💳 ${itemName}
-<br></br>
+&nbsp;
 | ⌘\t**Field** | **Value** |
 | --- | --- |
 ${Object.entries(card)
@@ -38,6 +38,7 @@ ${Object.entries(card)
 
     return `| ${shortcutKey ? `${shortcutKey}.` : "&nbsp;"}\t**${label}** | ${getCardValue(key, value)} |`;
   })
+  .filter(Boolean)
   .join("\n")}
 `;
 }