+++ extensions/github/src/components/IssueActions.tsx
@@ -224,7 +224,7 @@ export default function IssueActions({ issue, mutateList, mutateDetail, children
 
   const isAssignedToMe = issue.assignees.nodes?.some((assignee) => assignee?.isViewer);
 
-  const linkedBranch: any = issue.linkedBranches?.nodes?.length ? issue.linkedBranches.nodes[0] : null;
+  const linkedBranch = issue.linkedBranches?.nodes?.length ? issue.linkedBranches.nodes[0] : null;
 
   return (
     <ActionPanel title={`#${issue.number} in ${issue.repository.nameWithOwner}`}>
@@ -286,7 +286,7 @@ export default function IssueActions({ issue, mutateList, mutateDetail, children
                 title={"Delete Issue Branch"}
                 style={Action.Style.Destructive}
                 icon={{ source: "branch.svg", tintColor: Color.Red }}
-                onAction={() => deleteLinkedBranch(linkedBranch.id, linkedBranch.ref.name)}
+                onAction={() => deleteLinkedBranch((linkedBranch as any)?.id || "", linkedBranch.ref?.name ?? "")}
               />
             ) : null}
           </>
@@ -312,9 +312,9 @@ export default function IssueActions({ issue, mutateList, mutateDetail, children
           shortcut={{ modifiers: ["ctrl", "shift"], key: "," }}
         />
 
-        {linkedBranch ? (
+        {linkedBranch?.ref?.name ? (
           <Action.CopyToClipboard
-            content={linkedBranch.ref.name}
+            content={linkedBranch.ref?.name}
             title="Copy Branch Name"
             shortcut={{ modifiers: ["ctrl", "shift"], key: "." }}
           />
@@ -352,7 +352,7 @@ function AddAssigneeSubmenu({ issue, mutate }: SubmenuProps) {
       });
     },
     [issue],
-    { execute: load }
+    { execute: load },
   );
 
   async function addAssignee({ id, text }: { id: string; text: string }) {
@@ -428,7 +428,7 @@ function AddProjectSubmenu({ issue, mutate }: SubmenuProps) {
       });
     },
     [issue],
-    { execute: load }
+    { execute: load },
   );
 
   async function addProject({ id, text }: { id: string; text: string }) {
@@ -496,7 +496,7 @@ function SetMilestoneSubmenu({ issue, mutate }: SubmenuProps) {
       });
     },
     [issue],
-    { execute: load }
+    { execute: load },
   );
 
   async function unsetMilestone() {