+++ extensions/github/src/create-pull-request.tsx
@@ -59,7 +59,7 @@ export function PullRequestForm({ draftValues }: PullRequestFormProps) {
 
           // It's not possible to add a PR to a project from the initPullRequest call
           await Promise.all(
-            values.projects.map((projectId) => github.addPullRequestToProject({ pullRequestId, projectId }))
+            values.projects.map((projectId) => github.addPullRequestToProject({ pullRequestId, projectId })),
           );
 
           const pullRequest = updateResult?.updatePullRequest?.pullRequest;
@@ -134,7 +134,7 @@ export function PullRequestForm({ draftValues }: PullRequestFormProps) {
       return github.dataForRepository({ owner: selectedRepository.owner.login, name: selectedRepository.name });
     },
     [values.repository],
-    { execute: !!values.repository }
+    { execute: !!values.repository },
   );
 
   const defaultBranch = data?.repository?.defaultBranchRef;
@@ -154,10 +154,10 @@ export function PullRequestForm({ draftValues }: PullRequestFormProps) {
   const milestones = data?.repository?.milestones?.nodes;
 
   useEffect(() => {
-    const template = data?.repository?.pullRequestTemplateLowercase ?? data?.repository?.pullRequestTemplateUppercase;
+    const template = data?.repository?.pullRequestTemplates?.[0];
 
-    if (template && "text" in template && template.text && !values.description) {
-      setValue("description", template.text);
+    if (template && template.body && !values.description) {
+      setValue("description", template.body);
     }
 
     if (defaultBranch) {